gather util functions in utils.py
