+++ game/butterfly.py
@@ -10,28 +10,6 @@ from .entity import Entity
 from .util import *
 
 
-def rgb2hsv(r, g, b):
-    """Conversion between rgb in range 0-255 to hsv"""
-    return rgb_to_hsv(r / 255, g / 255, b / 255)
-
-
-def hsv2rgb(h, s, v):
-    """Conversion between hsv to rgb in range 0-255"""
-    s = clamp(s)
-    v = clamp(v)
-
-    r, g, b = hsv_to_rgb(h % 1, s, v)
-    return (
-        int(r * 255),
-        int(g * 255),
-        int(b * 255),
-    )
-
-
-def random_color():
-    """Random RGB color of the rainbow"""
-    return hsv2rgb(random(), 1, 1)
-
 
 class Butterfly(Entity):
     NB_FRAMES = 4
